addressing PR comments.
