+++ examples/air-purifier-app/air-purifier-common/include/air-purifier-manager.h
@@ -19,7 +19,6 @@
 #include <DeviceManager.h>
 #include <air-quality-sensor-manager.h>
 #include <app/clusters/fan-control-server/fan-control-server.h>
-#include <app/clusters/resource-monitoring-server/CodegenIntegration.h>
 #include <filter-delegates.h>
 #include <relative-humidity-sensor-manager.h>
 #include <temperature-sensor-manager.h>
