+++ scripts/tools/zap/tests/outputs/all-clusters-app/app-templates/CodeDrivenCallback.h
@@ -103,10 +103,6 @@ void MatterBooleanStateClusterInitCallback(chip::EndpointId endpointId);
 
 void MatterBooleanStateClusterShutdownCallback(chip::EndpointId endpointId);
 
-void MatterBooleanStateConfigurationClusterInitCallback(chip::EndpointId endpointId);
-
-void MatterBooleanStateConfigurationClusterShutdownCallback(chip::EndpointId endpointId);
-
 void MatterHepaFilterMonitoringClusterInitCallback(chip::EndpointId endpointId);
 
 void MatterHepaFilterMonitoringClusterShutdownCallback(chip::EndpointId endpointId);
@@ -114,3 +110,7 @@ void MatterHepaFilterMonitoringClusterShutdownCallback(chip::EndpointId endpoint
 void MatterActivatedCarbonFilterMonitoringClusterInitCallback(chip::EndpointId endpointId);
 
 void MatterActivatedCarbonFilterMonitoringClusterShutdownCallback(chip::EndpointId endpointId);
+
+void MatterBooleanStateConfigurationClusterInitCallback(chip::EndpointId endpointId);
+
+void MatterBooleanStateConfigurationClusterShutdownCallback(chip::EndpointId endpointId);
