+++ scripts/tools/zap/tests/outputs/all-clusters-app/app-templates/CodeDrivenInitShutdown.cpp
@@ -87,13 +87,14 @@ void MatterClusterServerInitCallback(EndpointId endpoint, ClusterId clusterId)
     case app::Clusters::BooleanState::Id:
         MatterBooleanStateClusterInitCallback(endpoint);
         break;
-    case app::Clusters::BooleanStateConfiguration::Id:
-        MatterBooleanStateConfigurationClusterInitCallback(endpoint);
     case app::Clusters::HepaFilterMonitoring::Id:
         MatterHepaFilterMonitoringClusterInitCallback(endpoint);
         break;
     case app::Clusters::ActivatedCarbonFilterMonitoring::Id:
         MatterActivatedCarbonFilterMonitoringClusterInitCallback(endpoint);
+        break;
+    case app::Clusters::BooleanStateConfiguration::Id:
+        MatterBooleanStateConfigurationClusterInitCallback(endpoint);
         break;
     }
 }
@@ -162,13 +163,14 @@ void MatterClusterServerShutdownCallback(EndpointId endpoint, ClusterId clusterI
     case app::Clusters::BooleanState::Id:
         MatterBooleanStateClusterShutdownCallback(endpoint);
         break;
-    case app::Clusters::BooleanStateConfiguration::Id:
-        MatterBooleanStateConfigurationClusterShutdownCallback(endpoint);
     case app::Clusters::HepaFilterMonitoring::Id:
         MatterHepaFilterMonitoringClusterShutdownCallback(endpoint);
         break;
     case app::Clusters::ActivatedCarbonFilterMonitoring::Id:
         MatterActivatedCarbonFilterMonitoringClusterShutdownCallback(endpoint);
+        break;
+    case app::Clusters::BooleanStateConfiguration::Id:
+        MatterBooleanStateConfigurationClusterShutdownCallback(endpoint);
         break;
     }
 }
