+++ src/app/clusters/resource-monitoring-server/CodegenIntegration.cpp
@@ -14,7 +14,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-#include "CodegenIntegration.h"
+#include "resource-monitoring-server.h"
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Attributes.h>
@@ -23,8 +23,6 @@
 #include <app/CommandHandler.h>
 #include <app/ConcreteCommandPath.h>
 #include <app/InteractionModelEngine.h>
-#include <app/clusters/identify-server/IdentifyCluster.h>
-#include <app/clusters/resource-monitoring-server/resource-monitoring-server.h>
 #include <app/server-cluster/DefaultServerCluster.h>
 #include <app/static-cluster-config/ActivatedCarbonFilterMonitoring.h>
 #include <app/static-cluster-config/HepaFilterMonitoring.h>
@@ -53,22 +51,22 @@ using namespace chip::app::Clusters::ResourceMonitoring;
 using chip::Protocols::InteractionModel::Status;
 
 // Helper function to register a legacy instance with the codegen data model provider
-inline void RegisterInstance(Instance * instance, EndpointId endpointId, ClusterId clusterId)
+void RegisterInstance(Instance * instance)
 {
     CHIP_ERROR err = CodegenDataModelProvider::Instance().Registry().Register(instance->mCluster.Registration());
 
     if (err != CHIP_NO_ERROR)
     {
 #if CHIP_CODEGEN_CONFIG_ENABLE_CODEGEN_INTEGRATION_LOOKUP_ERRORS
-        ChipLogError(AppServer, "Failed to register cluster %u/" ChipLogFormatMEI ":   %" CHIP_ERROR_FORMAT, endpointId,
-                     ChipLogValueMEI(instance->mCluster.Cluster().GetClusterId()), err.Format());
+        ChipLogError(AppServer, "Failed to register cluster %u/" ChipLogFormatMEI ":   %" CHIP_ERROR_FORMAT,
+                     instance->mCluster.Cluster().GetEndpointId(), ChipLogValueMEI(instance->mCluster.Cluster().GetClusterId()),
+                     err.Format());
 #endif // CHIP_CODEGEN_CONFIG_ENABLE_CODEGEN_INTEGRATION_LOOKUP_ERRORS
     }
 }
 
 // Helper function to construct Optional Attribute Set for ResourceMonitoring
-inline chip::Optional<ResourceMonitoringCluster::OptionalAttributeSet> ConstructOptionalAttributeSet(EndpointId endpointId,
-                                                                                                     ClusterId clusterId)
+ResourceMonitoringCluster::OptionalAttributeSet ConstructOptionalAttributeSet(EndpointId endpointId, ClusterId clusterId)
 {
     VerifyOrDie(emberAfContainsServer(endpointId, clusterId));
 
@@ -94,7 +92,7 @@ inline chip::Optional<ResourceMonitoringCluster::OptionalAttributeSet> Construct
         optionalAttributeSet.Set<ResourceMonitoring::Attributes::LastChangedTime::Id>();
     }
 
-    return chip::Optional<ResourceMonitoringCluster::OptionalAttributeSet>{ optionalAttributeSet };
+    return optionalAttributeSet;
 }
 
 } // namespace
@@ -109,13 +107,13 @@ Instance::Instance(Delegate * delegate, EndpointId endpointId, ClusterId cluster
     mCluster(endpointId, clusterId, BitFlags<ResourceMonitoring::Feature>{ featureMap },
              ConstructOptionalAttributeSet(endpointId, clusterId), degradationDirection, resetConditionCommandSupported)
 {
-    mCluster.Cluster().SetDelegate(delegate);
-    RegisterInstance(this, endpointId, clusterId);
+    TEMPORARY_RETURN_IGNORED mCluster.Cluster().SetDelegate(delegate);
+    RegisterInstance(this);
 }
 
 Instance::~Instance()
 {
-    CodegenDataModelProvider::Instance().Registry().Unregister(&mCluster.Cluster());
+    TEMPORARY_RETURN_IGNORED CodegenDataModelProvider::Instance().Registry().Unregister(&mCluster.Cluster());
 }
 
 } // namespace ResourceMonitoring
