+++ src/app/clusters/resource-monitoring-server/app_config_dependent_sources.gni
@@ -14,4 +14,5 @@
 app_config_dependent_sources = [
   "CodegenIntegration.cpp",
   "CodegenIntegration.h",
+  "resource-monitoring-server.h",
 ]
